refactor(view): WebViewMapper 제거로 인한 패키지 이동
